+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Grid options for the Section plugin
+- Make user waive its right of withdrawal when purchasing
+  a product with `is_withdrawable` set to `false`
 
 ## [2.30.0] - 2024-10-16